+++ example/big_date_example.dart
@@ -1,12 +1,11 @@
 import 'package:big_date/big_date.dart';
 
 void main() {
-  // DateTime d = DateTime(2022, 3, 2);
-  // d = d.toJapanDate();
-  // print(d.format("WW"));
-  // print(d.locatedWeekDay);
-
-  print(DateTime.now().lastMonth.lastMonth.lastMonth.dayCountOfMonth);
+  final now = DateTime.now();
+  print(now.lastMonth.lastMonth.lastMonth.dayCountOfMonth);
   DateTime.now().forEachDayOfMonthReverse((d) => print(d));
-  print((-3).abs() % 10);
+  print(now.format("%w"));
+  print(now.toJapanDate().format("%W"));
+  print(now.format("%_w"));
+  print(now.toJapanDate().format("%_W"));
 }